Splitting out the GetOrderDetails files
