+++ src/Web/Controllers/OrderController.cs
@@ -1,7 +1,6 @@
 ﻿using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.eShopWeb.Web.Features.Orders;
 using Microsoft.eShopWeb.Web.Features.Orders.Requests;
 using System.Threading.Tasks;
