+++ src/Web/Features/Orders/Requests/GetOrderDetails.cs
@@ -1,0 +1,17 @@
+﻿using MediatR;
+using Microsoft.eShopWeb.Web.ViewModels;
+
+namespace Microsoft.eShopWeb.Web.Features.Orders.Requests
+{
+    public class GetOrderDetails : IRequest<OrderViewModel>
+    {
+        public string UserName { get; set; }
+        public int OrderId { get; set; }
+
+        public GetOrderDetails(string userName, int orderId)
+        {
+            UserName = userName;
+            OrderId = orderId;
+        }
+    }
+}
