+++ tests/UnitTests/MediatorHandlers/OrdersTests/GetOrderDetails_Should.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.eShopWeb.ApplicationCore.Entities.OrderAggregate;
 using Microsoft.eShopWeb.ApplicationCore.Interfaces;
-using Microsoft.eShopWeb.Web.Features.Orders;
+using Microsoft.eShopWeb.Web.Features.Orders.RequestHandlers;
+using Microsoft.eShopWeb.Web.Features.Orders.Requests;
 using Moq;
 using System.Collections.Generic;
 using System.Threading;
